+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 [Downloadable releases](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- new method K.Util.setPath set sub property of object by path string
 - new method K.Util.json2html transform json or literal object into a nested ul list
 - plugin admin new settings adminsAutoFriendship
 - plugin geoinfo new method K.Geoinfo.getIpInfo
 - plugin admin new method for debug data in database, sanitizePlacesField
+- plugin geoinfo sanitize text values
+- plugin stats new stat findPlacesByField 
 
 ## v1.6.1
 - new user field loginIp